Map data type xsd:decimal to python Decimal

+++ tests/fixtures/defxmlschema/chapter15/example1508.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from dataclasses import dataclass, field
 from typing import Optional
 
@@ -16,7 +17,7 @@ class IdentifierGroup:
             required=True
         )
     )
-    version: Optional[float] = field(
+    version: Optional[Decimal] = field(
         default=None,
         metadata=dict(
             name="version",
